+++ Events.php
@@ -2,25 +2,15 @@
 
 namespace humhub\modules\fcmPush;
 
-use humhub\components\mail\Message;
-use humhub\helpers\DeviceDetectorHelper;
 use humhub\modules\fcmPush\assets\FcmPushAsset;
 use humhub\modules\fcmPush\assets\FirebaseAsset;
-use humhub\modules\fcmPush\components\MailerMessage;
 use humhub\modules\fcmPush\components\NotificationTargetProvider;
 use humhub\modules\fcmPush\helpers\MobileAppHelper;
 use humhub\modules\fcmPush\helpers\WebAppHelper;
 use humhub\modules\fcmPush\services\DriverService;
-use humhub\modules\fcmPush\widgets\PushNotificationInfoWidget;
 use humhub\modules\notification\targets\MobileTargetProvider;
-use humhub\modules\ui\menu\MenuLink;
-use humhub\modules\user\widgets\AccountTopMenu;
-use humhub\modules\user\widgets\AuthChoice;
-use humhub\modules\web\pwa\controllers\ManifestController;
 use humhub\modules\web\pwa\controllers\ServiceWorkerController;
-use humhub\widgets\BaseStack;
 use Yii;
-use yii\base\Event;
 
 class Events
 {
@@ -32,17 +22,6 @@ class Events
         if ($module->getDriverService()->hasConfiguredDriver()) {
             Yii::$container->set(MobileTargetProvider::class, NotificationTargetProvider::class);
         }
-
-        if ($module->getGoService()->isConfigured()) {
-            Yii::$container->set(Message::class, MailerMessage::class);
-        }
-    }
-
-    public static function onManifestControllerInit($event)
-    {
-        /** @var ManifestController $controller */
-        $controller = $event->sender;
-        $controller->manifest['gcm_sender_id'] = (string)103953800507;
     }
 
     public static function onServiceWorkerControllerInit($event): void
@@ -79,24 +58,8 @@ class Events
 JS;
     }
 
-    public static function onNotificationInfoWidget($event)
-    {
-        /** @var BaseStack $baseStack */
-        $baseStack = $event->sender;
-
-        $baseStack->addWidget(PushNotificationInfoWidget::class);
-    }
-
     public static function onLayoutAddonInit($event)
     {
-        // If the mobile app Opener page is open (after login and switching instance)
-        if (Yii::$app->session->has(MobileAppHelper::SESSION_VAR_HIDE_OPENER)) {
-            MobileAppHelper::registerHideOpenerScript();
-            Yii::$app->session->remove(MobileAppHelper::SESSION_VAR_HIDE_OPENER);
-        } elseif (DeviceDetectorHelper::appOpenerState()) {
-            MobileAppHelper::registerHideOpenerScript();
-        }
-
         // After login
         if (Yii::$app->session->has(MobileAppHelper::SESSION_VAR_REGISTER_NOTIFICATION)) {
             MobileAppHelper::registerNotificationScript();
@@ -113,13 +76,6 @@ JS;
             MobileAppHelper::unregisterNotificationScript();
             Yii::$app->session->remove(MobileAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION);
         }
-        if (Yii::$app->session->has(MobileAppHelper::SESSION_VAR_SHOW_OPENER)) {
-            MobileAppHelper::registerShowOpenerScript();
-            Yii::$app->session->remove(MobileAppHelper::SESSION_VAR_SHOW_OPENER);
-        }
-
-        // Get info for the Share intend feature (uploading files from the mobile app)
-        MobileAppHelper::getFileUploadSettings();
 
         if (!Yii::$app->user->isGuest) {
             static::registerAssets();
@@ -141,7 +97,6 @@ JS;
 
     public static function onAfterLogin()
     {
-        Yii::$app->session->set(MobileAppHelper::SESSION_VAR_HIDE_OPENER, 1);
         Yii::$app->session->set(MobileAppHelper::SESSION_VAR_REGISTER_NOTIFICATION, 1);
     }
 
@@ -149,39 +104,5 @@ JS;
     {
         Yii::$app->session->set(WebAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION, 1);
         Yii::$app->session->set(MobileAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION, 1);
-        if (!DeviceDetectorHelper::isAppRequest()) {
-            Yii::$app->session->set(MobileAppHelper::SESSION_VAR_SHOW_OPENER, 1);
-        }
-    }
-
-    public static function onAuthChoiceBeforeRun(Event $event)
-    {
-        /** @var AuthChoice $sender */
-        $sender = $event->sender;
-
-        /** @var Module $module */
-        $module = Yii::$app->getModule('fcm-push');
-
-        if (DeviceDetectorHelper::isIosApp() && $module->getConfigureForm()->disableAuthChoicesIos) {
-            $sender->setClients([]);
-        }
-    }
-
-    public static function onAccountTopMenuInit(Event $event)
-    {
-        if (!DeviceDetectorHelper::isMultiInstanceApp()) {
-            return;
-        }
-
-        /** @var AccountTopMenu $menu */
-        $menu = $event->sender;
-
-        $menu->addEntry(new MenuLink([
-            'label' => Yii::t('FcmPushModule.base', 'Switch network'),
-            'url' => ['/fcm-push/mobile-app/instance-opener'],
-            'icon' => 'arrows-h',
-            'sortOrder' => 699, // Just before "Logout"
-            'isVisible' => true,
-        ]));
     }
 }